refactor(user): Split errors, policies, and schemas
